exclude folders from sync (#16)
Including the environment variable foldersToSync was originally added as a workaround to avoid syncing the ‘Synapse user ID’ folders (those with the 7-digit format e.g. ‘3413795’) to Synapse. When filehandles are added to Synapse, some metadata(I believe?)/auxiliary files are created by Synapse, which are then stored in the bucket under this new folder. These files will then trigger the lambda, and without some restriction, filehandles will be created for those files and start a loop.

A member of the DCC did a test run of the data ingress process last week and expressed that it will be inconvenient to ask centers to pre-plan which top-level folders they will be using, and to have them inform us of these folders/any subsequent changes. We are thinking of using this somewhat hacky solution and instead of specifying which folders to sync, we will exclude those ’Synapse folders’ by only syncing folders that do not begin with a number (and instruct centers to not use such folders). This will hopefully allow centers more flexibility to add/change folders as they need.

* exclude folders from sync
* update lambda to use env vars directly